requirements.txt
# API Testing Framework - Python Requirements

# Core dependencies
requests>=2.28.1
pytest>=7.1.3
pytest-cov>=4.0.0
assertpy>=1.1
jsonschema>=4.17.0
pyjwt>=2.6.0
python-dotenv>=0.21.0
allure-pytest>=2.12.0

# Performance testing
locust>=2.12.1

# Security testing
owasp-zap-api-python>=0.4.0
bandit>=1.7.4

# Utility libraries
pyyaml>=6.0
deepdiff>=6.2.1

# Logging
loguru>=0.6.0
